feat: created the component button
