+++ SiteSherchingSolutions/package.json
@@ -14,6 +14,7 @@
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.4",
     "react-dom": "^18.3.1",
+    "react-icons": "^5.3.0",
     "react-router-dom": "^6.26.2",
     "sitesherchingsolutions": "file:"
   },
