+++ SiteSherchingSolutions/src/Pages/Home.jsx
@@ -1,5 +1,5 @@
 import Slogan from '../componentes/Header/Slogan'
-import NavBar from '../layout/NavBar'
+import NavBar from '../componentes/navBar/NavBar'
 import style from './Home.module.css'
 
 
@@ -7,16 +7,9 @@ export function Home() {
   
 
   return (
-    <div className={`${style.home}`}>
-      <nav>
+    <div className={style.home}>
         <NavBar/>
-      </nav>
-      <section>
         <Slogan/>
-      </section>
-        
-      
-       
     </div>
    
   )
