+++ SiteSherchingSolutions/src/Pages/Home.module.css
@@ -6,5 +6,4 @@
     width: 100%;
     height: 100vh;
     background-color: #010035;
-    ;
 }
