+++ SiteSherchingSolutions/src/componentes/Header/Slogan.module.css
@@ -1,28 +1,41 @@
-.slogan{
-    color: #ffffff;
+.slogan {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
 }
 
-
-.textPrimary{
-    margin-top: 12.5rem;
+.textPrimary {
     font-size: 3rem;
     text-decoration: none;
     font-weight: 200;
+    line-height:1.2;
+    max-width: 763.6px;
+    margin-top: 12.5rem;
 }
 
-.slogan span{
+.slogan span {
+    margin-left: 0.75rem;
     color:#0017E8;
     font-weight: 300;
+    position: relative;
 }
 
-.vector{
-    margin-top: -5.2rem;
-    padding-left: 21rem;
-    height: 0.9rem;
+.vector {
+    width: 100%;
+    position: absolute;
+    right: 0px;
+    bottom: 0;
 }
 
-.textSegundary{
+.textSegundary {
     font-weight: 200;
     font-size: 2rem;
     line-height: 1;
+    max-width: 763px;
+    
+}
+
+.btnContato {
+    background-color: black;
+    color: #ffffff;
 }
