+++ SiteSherchingSolutions/src/index.css
@@ -1,15 +1,5 @@
 body{
   background-color:  #010035;
-  
-}
-h1{
-  font-family: "Poppins", sans-serif;
-}
-a{
-  font-family: "Poppins", sans-serif;
-}
-
-p{
-  font-family: "Poppins";
-  color: #ffffff;
+  font-family: "Poppins" !important; 
+  color: #ffffff !important;
 }
