+++ SiteSherchingSolutions/src/layout/Button.jsx
@@ -1,0 +1,14 @@
+import style from './Button.module.css'
+
+
+
+// eslint-disable-next-line react/prop-types
+function Button({onClick, className, children}) {
+    return (
+        <button onClick={onClick} className={`${style.btnComponente} ${className}`}>
+            {children}
+        </button>
+    )
+}
+
+export default Button
