+++ SiteSherchingSolutions/src/layout/Button.module.css
@@ -1,0 +1,6 @@
+.btnComponente{
+    border: none;
+    border-radius: 1.75rem;
+    height: 2.5rem;
+    width: 14.81rem;
+}
