+++ SiteSherchingSolutions/src/router.jsx
@@ -5,7 +5,7 @@ const Router = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route index element={<Home />}></Route>
+        <Route index element={<Home />}/>
       </Routes>
     </BrowserRouter>
   );
