--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "google-fonts": "^1.0.0"
-  }
-}
